+++ apps/cli/clixon_cli_api.h
@@ -117,6 +117,7 @@ int cli_help(clixon_handle h, cvec *vars, cvec *argv);
 cvec *cvec_append(cvec *cvv0, cvec *cvv1);
 int   cvec_concat_cb(cvec *cvv, cbuf *cb);
 int cli_process_control(clixon_handle h, cvec *vars, cvec *argv);
+int cli_alias_cb(clixon_handle h, cvec *cvv, cvec *argv);
 
 /* In cli_show.c */
 int expand_dbvar(void *h, char *name, cvec *cvv, cvec *argv,